+++ modules/ensemble/lib/framework/action.dart
@@ -21,6 +21,7 @@ import 'package:ensemble/action/notification_actions.dart';
 import 'package:ensemble/action/phone_contact_action.dart';
 import 'package:ensemble/action/sign_in_out_action.dart';
 import 'package:ensemble/action/toast_actions.dart';
+import 'package:ensemble/action/disable_hardware_navigation.dart';
 import 'package:ensemble/ensemble.dart';
 import 'package:ensemble/framework/data_context.dart';
 import 'package:ensemble/framework/error_handling.dart';
@@ -49,12 +50,16 @@ class ShowCameraAction extends EnsembleAction {
     this.onComplete,
     this.onClose,
     this.onCapture,
+    this.onError,
+    this.overlayWidget,
   }) : super(initiator: initiator);
   final Map<String, dynamic>? options;
   String? id;
   EnsembleAction? onComplete;
   EnsembleAction? onClose;
   EnsembleAction? onCapture;
+  EnsembleAction? onError;
+  dynamic overlayWidget;
 
   factory ShowCameraAction.fromYaml({Invokable? initiator, Map? payload}) {
     return ShowCameraAction(
@@ -64,6 +69,8 @@ class ShowCameraAction extends EnsembleAction {
       onComplete: EnsembleAction.from(payload?['onComplete']),
       onClose: EnsembleAction.from(payload?['onClose']),
       onCapture: EnsembleAction.from(payload?['onCapture']),
+      onError: EnsembleAction.from(payload?['onError']),
+      overlayWidget: payload?['overlayWidget'],
     );
   }
 }
@@ -1048,6 +1055,7 @@ enum ActionType {
   bluetoothDisconnect,
   bluetoothSubscribeCharacteristic,
   bluetoothUnsubscribeCharacteristic,
+  controlDeviceBackNavigation
 }
 
 /// payload representing an Action to do (navigateToScreen, InvokeAPI, ..)
@@ -1167,6 +1175,8 @@ abstract class EnsembleAction {
       return CopyToClipboardAction.from(payload: payload);
     } else if (actionType == ActionType.share) {
       return ShareAction.from(payload: payload);
+    } else if (actionType == ActionType.controlDeviceBackNavigation) {
+      return ControlBackNavigation.from(payload: payload);
     } else if (actionType == ActionType.rateApp) {
       return RateAppAction.from(payload: payload);
     } else if (actionType == ActionType.getDeviceToken) {